[REFACTOR] uploader 모듈화, /test 컨트롤러 주석달기
